requirements.txt
openai
pinecone
python-dotenv
tqdm
fastapi
uvicorn
